docker/patches/requirements-cache.txt
# Gemini Prompt Caching - OpenTelemetry/Prometheus Metrics Dependencies
# Feature: 006-gemini-prompt-caching
# Purpose: Expose cache statistics via Prometheus metrics endpoint

# OpenTelemetry SDK for metrics instrumentation
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Prometheus exporter for OpenTelemetry
opentelemetry-exporter-prometheus>=0.41b0

# Prometheus client library
prometheus-client>=0.17.0
